# Refactor checkOrphanPages.ts