I nerd-sniped myself into this change. I was trying to improve the performance of the orphan check for https://github.com/Qiskit/documentation/issues/2800 by using `Promises.all()`, but it turns out it barely makes an improvement.

This PR:

```
❯ hyperfine "npx tsx scripts/js/commands/checkOrphanPages.ts -- --apis" 
Benchmark 1: npx tsx scripts/js/commands/checkOrphanPages.ts -- --apis
  Time (mean ± σ):     587.6 ms ±  14.6 ms    [User: 778.5 ms, System: 155.7 ms]
  Range (min … max):   567.8 ms … 621.2 ms    10 runs
```

Main branch:

```
❯ hyperfine "npx tsx scripts/js/commands/checkOrphanPages.ts -- --apis"
Benchmark 1: npx tsx scripts/js/commands/checkOrphanPages.ts -- --apis
  Time (mean ± σ):     592.6 ms ±  18.8 ms    [User: 781.1 ms, System: 153.7 ms]
  Range (min … max):   570.7 ms … 631.7 ms    10 runs
```

So, this PR solely refactors the code:

* Uses `Promises.all` with `.map()` for more idiomatic code. (The parallization doesn't impact performance much)
* Improves the error message
* Generally uses more concise code, like `jsonFileContents` -> `raw`
* Uses more consistent naming:
    -`findX` means find file paths in file system
    - `readX` means read file X
    - `parseX` means convert a string to something parsed